delete process model
