+++ dinky-admin/src/main/java/org/dinky/utils/MavenUtil.java
@@ -19,9 +19,9 @@
 
 package org.dinky.utils;
 
+import org.dinky.data.exception.DinkyException;
 import org.dinky.data.model.SystemConfiguration;
 import org.dinky.function.constant.PathConstant;
-import org.dinky.process.exception.DinkyException;
 
 import java.io.File;
 import java.io.IOException;
