+++ dinky-common/src/main/java/org/dinky/data/annotations/ExecuteProcess.java
@@ -17,25 +17,20 @@
  *
  */
 
-package org.dinky.process.exception;
+package org.dinky.data.annotations;
 
-public class ExcuteException extends Exception {
+import org.dinky.data.enums.ProcessType;
 
-    public ExcuteException() {}
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
-    public ExcuteException(String message) {
-        super(message);
-    }
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+@Inherited
+public @interface ExecuteProcess {
 
-    public ExcuteException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ExcuteException(Throwable cause) {
-        super(cause);
-    }
-
-    public ExcuteException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
+    ProcessType type();
 }
