+++ dinky-common/src/main/java/org/dinky/data/annotations/ProcessId.java
@@ -17,9 +17,7 @@
  *
  */
 
-package org.dinky.process.annotations;
-
-import org.dinky.process.enums.ProcessStepType;
+package org.dinky.data.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
@@ -27,9 +25,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
-public @interface ProcessStep {
-    ProcessStepType type();
-}
+public @interface ProcessId {}
