+++ dinky-core/src/main/java/org/dinky/job/JobManager.java
@@ -29,6 +29,9 @@ import org.dinky.constant.FlinkSQLConstant;
 import org.dinky.context.CustomTableEnvironmentContext;
 import org.dinky.context.FlinkUdfPathContextHolder;
 import org.dinky.context.RowLevelPermissionsContext;
+import org.dinky.data.annotations.ProcessStep;
+import org.dinky.data.enums.ProcessStepType;
+import org.dinky.data.exception.DinkyException;
 import org.dinky.data.model.FlinkUdfManifest;
 import org.dinky.data.model.SystemConfiguration;
 import org.dinky.data.result.ErrorResult;
@@ -58,9 +61,6 @@ import org.dinky.interceptor.FlinkInterceptor;
 import org.dinky.interceptor.FlinkInterceptorResult;
 import org.dinky.parser.SqlType;
 import org.dinky.parser.check.AddJarSqlParser;
-import org.dinky.process.annotations.ProcessStep;
-import org.dinky.process.enums.ProcessStepType;
-import org.dinky.process.exception.DinkyException;
 import org.dinky.trans.ExecuteJarParseStrategy;
 import org.dinky.trans.Operations;
 import org.dinky.trans.dml.ExecuteJarOperation;
