+++ dinky-gateway/pom.xml
@@ -47,10 +47,6 @@
         </dependency>
         <dependency>
             <groupId>org.dinky</groupId>
-            <artifactId>dinky-process</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.dinky</groupId>
             <artifactId>dinky-client-${dinky.flink.version}</artifactId>
             <scope>${scope.runtime}</scope>
         </dependency>
