+++ dinky-metadata/dinky-metadata-base/pom.xml
@@ -32,10 +32,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.dinky</groupId>
-            <artifactId>dinky-process</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.alibaba</groupId>
             <artifactId>druid-spring-boot-starter</artifactId>
         </dependency>
