+++ dinky-metadata/pom.xml
@@ -49,6 +49,10 @@
             <groupId>com.github.xiaoymin</groupId>
             <artifactId>knife4j-openapi2-spring-boot-starter</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.dinky</groupId>
+            <artifactId>dinky-common</artifactId>
+        </dependency>
     </dependencies>
 
 </project>
